\name{odm.close}
\alias{odm.close}
\title{
Close existing connection to observations database.
}
\description{
Close the existing connection to observations database.
}
\usage{
odm.close()
}

\value{
Nothing returned.
}
\author{
Dominik Reusser
}
\examples{
getMetadata(table="VariableName")
odm.close()
}
\keyword{ utilities }
